tests/wpt/web-platform-tests/content-security-policy/blink-contrib-2/base-uri-allow.sub.html

<!DOCTYPE html>
<html>

<head>
    <!-- Programmatically converted from a WebKit Reftest, please forgive resulting idiosyncracies.-->
    <title>base-uri-allow</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src='../support/logTest.sub.js?logs=["TEST COMPLETE"]'></script>
    <script src="../support/alertAssert.sub.js?alerts=[]"></script>
    <!-- enforcing policy:
base-uri http://www1.{{host}}:{{ports[http][0]}}; script-src 'self' 'unsafe-inline' http://www1.{{host}}:{{ports[http][0]}}; connect-src 'self';
-->
    <base href="http://www1.{{host}}:{{ports[http][0]}}/">
    <script>
        console.info('Check that base URIs can be set if they do not violate the page\'s policy.');
        test(function() {
            assert_equals(document.baseURI, 'http://www1.{{host}}:{{ports[http][0]}}/');
            log("TEST COMPLETE")
        });

    </script>
</head>

<body>
    <div id="log"></div>
    <script async defer src="./content-security-policy/support/checkReport.sub.js?reportExists=false"></script>
</body>

</html>
